requirements.txt
flask
python-dotenv
openai
gunicorn

